# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jeigo API",
        "description": "Dictionary API",
        "version": "1.0.0"
    },
    "paths": {
...
        "user/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user details",
                "description": "Returns details about the currently authenticated user.",
                "operationId": "getAuthenticatedUser",
                "responses": {
                    "200": {
                        "description": "User details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "9e0896b5-8e6f-45b0-ad81-2d1d03e70a2b"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "john@testo.com"
                                        },
                                        "email_verified_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-23T01:54:14.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-23T01:54:14.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "user/me/favorites": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List favorite words",
                "operationId": "45d832760e48e351e04c2cf565b4f622",
                "parameters": [
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Cursor for pagination",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "word": {
                                                        "type": "string",
                                                        "example": "aahed"
                                                    },
                                                    "added": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-01-22T22:59:15.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost/api/user/me/favorites"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "next_cursor": {
                                            "type": "string",
                                            "example": "eyJ3b3Jkcy5pZCI6OCwiX3BvaW50c1RvTmV4dEl0ZW1zIjp0cnVlfQ"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/user/me/favorites?cursor=eyJ3b3Jkcy5pZCI6OCwiX3BvaW50c1RvTmV4dEl0ZW1zIjp0cnVlfQ"
                                        },
                                        "prev_cursor": {
                                            "type": "string",
                                            "example": "eyJ3b3Jkcy5pZCI6NSwiX3BvaW50c1RvTmV4dEl0ZW1zIjpmYWxzZX0"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/user/me/favorites?cursor=eyJ3b3Jkcy5pZCI6NSwiX3BvaW50c1RvTmV4dEl0ZW1zIjpmYWxzZX0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "entries/en/{word}/favorite": {
            "post": {
                "tags": [
                    "Words"
                ],
                "summary": "Favorite a word",
                "operationId": "c696bba311701d3f12798665834e4678",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "description": "The word to favorite",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "peach"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content (word favorited successfully)"
                    },
                    "404": {
                        "description": "Word not found"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "entries/en/{word}/unfavorite": {
            "delete": {
                "tags": [
                    "Words"
                ],
                "summary": "Unfavorite a word",
                "operationId": "611336cff5225511dab24c58c37a0d95",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "description": "The the word to unfavorite",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "peach"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content (word unfavorited successfully)"
                    },
                    "404": {
                        "description": "Word not found"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "user/me/history": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List words history",
                "operationId": "3a380f7d921c924ac2b23dc202d1a67a",
                "parameters": [
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Cursor for pagination",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "word": {
                                                        "type": "string",
                                                        "example": "aah"
                                                    },
                                                    "added": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-01-23T11:52:37.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost/api/user/me/history?cursor=eyJ3b3Jkcy5pZCI6MywiX3BvaW50c1RvTmV4dEl0ZW1zIjp0cnVlfQ"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "next_cursor": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "prev_cursor": {
                                            "type": "string",
                                            "example": "eyJ3b3Jkcy5pZCI6NCwiX3BvaW50c1RvTmV4dEl0ZW1zIjpmYWxzZX0"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/user/me/history?cursor=eyJ3b3Jkcy5pZCI6NCwiX3BvaW50c1RvTmV4dEl0ZW1zIjpmYWxzZX0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}